DoubleKingPedro v0.0.1 DoubleKingPedro.Card

Represents a deck of cards

Types

suite()

suite() :: :hearts | :diamonds | :clubs | :spades | :joker

Functions

deal(deck, number, hand \\ [])

deal(deck :: [%DoubleKingPedro.Card{suite: term, value: term}], number :: integer, hand :: [%DoubleKingPedro.Card{suite: term, value: term}]) :: {hand :: [%DoubleKingPedro.Card{suite: term, value: term}], deck :: [%DoubleKingPedro.Card{suite: term, value: term}]}

Deals cards out of cards to a hand (either empty or a list of %Card)

deck()

deck() :: [%DoubleKingPedro.Card{suite: term, value: term}, ...]

Constructs a new, unshuffled, deck

gen_suite(suite, value \\ 2)

gen_suite(suite, where_to_start :: integer) :: [%DoubleKingPedro.Card{suite: term, value: term}]

Generates a single suite of cards

same_color(suite1, suite2)

same_color(suite, suite) :: boolean

If the two given suites are of the same color, returns true, else, false.