DoubleKingPedro v0.0.1 DoubleKingPedro.Game

The Game module provides the logic for Double King Pedro

Types

gameState()

gameState ::
  :lobby |
  :bidding |
  :trump_select |
  :card_select |
  :tricks

Functions

join_game(game, player)

join_game(%DoubleKingPedro.Game{players: term, state: term, state_content: term, teams: term}, %DoubleKingPedro.Player{hand: term, id: term, name: term, position: term, team: term}) ::
  %DoubleKingPedro.Game{players: term, state: term, state_content: term, teams: term} |
  :game_full

Adds a %Player{} to the provided game; if the game is full, returns :game_full

make_move(game, player, move)

make_move(%DoubleKingPedro.Game{players: term, state: term, state_content: term, teams: term}, String.t, any) ::
  %DoubleKingPedro.Game{players: term, state: term, state_content: term, teams: term} |
  :not_your_turn |
  :invalid_move

The primary interface to the game; after the game has been started, all moves get sent to this function until the game ends.

It accepts the following forms:

make_move(game, player, player_bid) - where bid is a number between 0 and 100

make_move(game, player, trump) - where trump is one of :hearts, :diamonds, :spades, or :clubs

make_move(game, player, {action, card}) - where action is either :drop or :pass, and card is a card in the players hand

make_mopve(game, player, %Card{}) - players a card from the players hand

Each of these states can return: %Game (an updated game object), :not_your_turn, or :invalid_move. In general, :invalid_move means the player tried to something that wasn’t valid, such as playering a non-trump card during a trick where trump is required, and :not_your_turn means the player attempted to make a move when it wasn’t their turn.

If a player who is not in the game attempts to make a move, it returns :not_your_turn

start(game)

If the game is in a valid state to start playing, transitions to the playering state